+++ timelock-agent/src/main/java/com/palantir/atlasdb/timelock/paxos/SingleLeaderNetworkClientFactories.java
@@ -36,7 +36,7 @@ abstract class SingleLeaderNetworkClientFactories {
 
     abstract PaxosUseCase useCase();
     abstract TimelockPaxosMetrics metrics();
-    abstract ClientPaxosResourceFactory.PaxosRemoteClients remoteClients();
+    abstract PaxosRemoteClients remoteClients();
     abstract PaxosComponents components();
     abstract int quorumSize();
     abstract ExecutorService sharedExecutor();
